requirements-with-cupy.txt
torch
numpy
einops
opencv-contrib-python
kornia
scipy
Pillow
torchvision
tqdm
# cupy-wheel
cupy-cuda12x>=13.0.0
